Multithreaded

+++ Shifter.h
@@ -8,7 +8,7 @@ class Shifter {
 public:
     Shifter(double freq, double hz);
     Shifter();
-    void shift(std::complex<float>* x, size_t len, size_t start);
+    void shift(std::complex<float>* x, size_t len, uint64_t start);
 
 private:
     std::vector<std::complex<float>> shift_buff;
